ddtrace/tracer: switch atomics to 32-bit (#1443)
sync/atomic has several issues. Among them is that it causes a panic when
a 64-bit field isn't correctly aligned. Alignment must be manually
ensured and is easy to forget.

Instead, we will use 32-bit atomic integers which do not require manual
alignment. We can eventually trade them out for Go's new atomics APIs
that were introduced in go1.19, but we have to wait until 1.18 falls out
of our supported versions.

Fixes #1418

+++ ddtrace/tracer/tracer_test.go
@@ -353,8 +353,8 @@ func TestSamplingDecision(t *testing.T) {
 		tracer, _, _, stop := startTestTracer(t)
 		defer func() {
 			// Must check these after tracer is stopped to avoid flakiness
-			assert.Equal(t, uint64(1), tracer.droppedP0Traces)
-			assert.Equal(t, uint64(2), tracer.droppedP0Spans)
+			assert.Equal(t, uint32(1), tracer.droppedP0Traces)
+			assert.Equal(t, uint32(2), tracer.droppedP0Spans)
 		}()
 		defer stop()
 		tracer.config.agent.DropP0s = true
